+++ src/api/integration/integration_test.go
@@ -3473,137 +3473,6 @@ func TestLivePendingTransactionsVerbose(t *testing.T) {
 	require.NoError(t, err)
 }
 
-func TestLiveWalletSpend(t *testing.T) {
-	if !doLive(t) {
-		return
-	}
-
-	if liveDisableNetworking(t) {
-		t.Skip("Spend tests require networking")
-		return
-	}
-
-	requireWalletEnv(t)
-
-	c := newClient()
-	w, totalCoins, _, password := prepareAndCheckWallet(t, c, 2e6, 2)
-
-	tt := []struct {
-		name     string
-		to       string
-		coins    uint64
-		checkTxn func(t *testing.T, tx *readable.TransactionWithStatus)
-	}{
-		{
-			name:  "send all coins to the first address",
-			to:    w.Entries[0].Address.String(),
-			coins: totalCoins,
-			checkTxn: func(t *testing.T, tx *readable.TransactionWithStatus) {
-				// Confirms the total output coins are equal to the totalCoins
-				var coins uint64
-				for _, o := range tx.Transaction.Out {
-					c, err := droplet.FromString(o.Coins)
-					require.NoError(t, err)
-					coins, err = coin.AddUint64(coins, c)
-					require.NoError(t, err)
-				}
-
-				// Confirms the address balance are equal to the totalCoins
-				coins, _ = getAddressBalance(t, c, w.Entries[0].Address.String())
-				require.Equal(t, totalCoins, coins)
-			},
-		},
-		{
-			// send 0.003 coin to the second address,
-			// this amount is chosen to not interfere with TestLiveWalletCreateTransaction
-			name:  "send 0.003 coin to second address",
-			to:    w.Entries[1].Address.String(),
-			coins: 3e3,
-			checkTxn: func(t *testing.T, tx *readable.TransactionWithStatus) {
-				// Confirms there're two outputs, one to the second address, one as change output to the first address.
-				require.Len(t, tx.Transaction.Out, 2)
-
-				// Gets the output of the second address in the transaction
-				getAddrOutputInTxn := func(t *testing.T, tx *readable.TransactionWithStatus, addr string) *readable.TransactionOutput {
-					for _, output := range tx.Transaction.Out {
-						if output.Address == addr {
-							return &output
-						}
-					}
-					t.Fatalf("transaction doesn't have output to address: %v", addr)
-					return nil
-				}
-
-				out := getAddrOutputInTxn(t, tx, w.Entries[1].Address.String())
-
-				// Confirms the second address has 0.003 coin
-				require.Equal(t, out.Coins, "0.003000")
-				require.Equal(t, out.Address, w.Entries[1].Address.String())
-
-				coin, err := droplet.FromString(out.Coins)
-				require.NoError(t, err)
-
-				// Gets the expected change coins
-				expectChangeCoins := totalCoins - coin
-
-				// Gets the real change coins
-				changeOut := getAddrOutputInTxn(t, tx, w.Entries[0].Address.String())
-				changeCoins, err := droplet.FromString(changeOut.Coins)
-				require.NoError(t, err)
-				// Confirms the change coins are matched.
-				require.Equal(t, expectChangeCoins, changeCoins)
-			},
-		},
-	}
-
-	for _, tc := range tt {
-		t.Run(tc.name, func(t *testing.T) {
-			result, err := c.Spend(w.Filename(), tc.to, tc.coins, password)
-			if err != nil {
-				t.Fatalf("spend failed: %v", err)
-			}
-
-			tk := time.NewTicker(time.Second)
-			var txn *readable.TransactionWithStatus
-		loop:
-			for {
-				select {
-				case <-time.After(30 * time.Second):
-					t.Fatal("Waiting for transaction to be confirmed timeout")
-				case <-tk.C:
-					txn = getTransaction(t, c, result.Transaction.Hash)
-					if txn.Status.Confirmed {
-						break loop
-					}
-				}
-			}
-			tc.checkTxn(t, txn)
-		})
-	}
-
-	// Return if wallet is encrypted, cause the rest of the tests will spend a lot of time.
-	if w.IsEncrypted() {
-		return
-	}
-
-	// Confirms sending coins less than 0.001 is not allowed
-	errMsg := "500 Internal Server Error - Transaction violates soft constraint: invalid amount, too many decimal places"
-	for i := uint64(1); i < uint64(1000); i++ {
-		cs, err := droplet.ToString(i)
-		require.NoError(t, err)
-		name := fmt.Sprintf("send invalid coin %v", cs)
-		t.Run(name, func(t *testing.T) {
-			result, err := c.Spend(w.Filename(), w.Entries[0].Address.String(), i, password)
-			if w.IsEncrypted() && len(password) == 0 {
-				assertResponseError(t, err, http.StatusBadRequest, "400 Bad Request - missing password")
-				return
-			}
-			assertResponseError(t, err, http.StatusInternalServerError, errMsg)
-			require.Nil(t, result)
-		})
-	}
-}
-
 func TestStableInjectTransaction(t *testing.T) {
 	if !doStable(t) {
 		return
@@ -4100,7 +3969,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					Hours:   1,
 				},
 				{
-					// Address omitted -- will be check later in the test body
+					// Address omitted -- will be checked later in the test body
 					Coins: 1e3,
 					Hours: remainingHours - 1,
 				},
@@ -5729,20 +5598,6 @@ func TestDisableWalletAPI(t *testing.T) {
 			name:      "get wallet balance",
 			method:    http.MethodGet,
 			endpoint:  "/api/v1/wallet/balance?id=test.wlt",
-			expectErr: "403 Forbidden - Endpoint is disabled",
-			code:      http.StatusForbidden,
-		},
-		{
-			name:     "wallet spending",
-			method:   http.MethodPost,
-			endpoint: "/api/v1/wallet/spend",
-			body: func() io.Reader {
-				v := url.Values{}
-				v.Add("id", "test.wlt")
-				v.Add("coins", "100000") // 1e5
-				v.Add("dst", "2jBbGxZRGoQG1mqhPBnXnLTxK6oxsTf8os6")
-				return strings.NewReader(v.Encode())
-			},
 			expectErr: "403 Forbidden - Endpoint is disabled",
 			code:      http.StatusForbidden,
 		},